fix(Interactions): interactor to emit collision stopped on disable
There was an issue where an Interactor would not emit the
CollisionStopped event if the Interactor was disabled. This was due
to the internal Interactor Rule becoming disabled and rules would
always be rejected on disabled components.

A recent change in Zinnia now means the Rule can still be queried
even if the component is on a deactivated GameObject. This fix simply
allows the Rule within the Interactor to still be queried when the
Interactor is disabled and therefore the CollisionStopped event still
correctly emits.