feat(.github): use organization .github repository
GitHub provides a mechanism where a global organization .github repo
can be used as a fallback to provide default community health files
instead of repeating the same files across multiple repos.

ExtendRealityLtd now has a `.github` repo which should be used as
it provides the correct details for this repo.

The README.md has been updated to provide definitive links to the
relevant files.